# api/openapi.yaml
openapi: 3.1.0
info:
  title: Food History API
  version: 1.2.0
  description: >-
    Phase 1A-C API contract for the Food History knowledge base, menu archive,
    bibliographic catalogue, material-culture catalogue, and evidence graph.
servers:
  - url: https://api.example.org/v1
paths:
...
  /wiki/pages:
    get:
      summary: List published wiki articles
      parameters:
        - {name: q, in: query, schema: {type: string}, description: Full-text article search}
        - {name: category, in: query, schema: {type: string}, description: Category slug}
      responses:
        '200':
          description: Published articles
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/WikiPageSummary'}}}}
  /wiki/pages/{slug}:
    get:
      summary: Get the published revision of a wiki article
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Published wiki article
          content: {application/json: {schema: {$ref: '#/components/schemas/WikiPage'}}}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    staffToken: {type: http, scheme: bearer}
...
  responses:
    NotFound: {description: Resource not found}
  schemas:
    WikiPageSummary:
      type: object
      required: [public_id, slug, title]
      properties:
        public_id: {type: string}
        slug: {type: string}
        title: {type: string}
        summary: {type: [string, 'null']}
    WikiCitation:
      type: object
      required: [sequence, source_public_id, source_label]
      properties:
        sequence: {type: integer}
        source_public_id: {type: string}
        source_label: {type: string}
        locator_text: {type: [string, 'null']}
        page_label: {type: [string, 'null']}
        stable_uri: {type: [string, 'null'], format: uri}
        note: {type: [string, 'null']}
    WikiPage:
      allOf:
        - {$ref: '#/components/schemas/WikiPageSummary'}
        - type: object
          required: [revision_number, body_html, toc, backlinks, entities, citations, categories, images]
          properties:
            revision_number: {type: integer, minimum: 1}
            body_html: {type: string, description: Sanitized server-rendered article HTML}
            toc:
              type: array
              items:
                type: object
                required: [id, title, level]
                properties:
                  id: {type: string}
                  title: {type: string}
                  level: {type: integer, minimum: 2, maximum: 4}
            backlinks:
              type: array
              items: {$ref: '#/components/schemas/WikiPageSummary'}
            entities:
              type: array
              items:
                allOf:
                  - {$ref: '#/components/schemas/EntityRef'}
                  - type: object
                    properties: {relationship: {type: string, enum: [subject, mentions]}}
            citations: {type: array, items: {$ref: '#/components/schemas/WikiCitation'}}
            categories:
              type: array
              items:
                type: object
                required: [slug, name]
                properties:
                  slug: {type: string}
                  name: {type: string}
            images:
              type: array
              items:
                type: object
                required: [public_id, storage_uri, placement, alt_text]
                properties:
                  public_id: {type: string}
                  storage_uri: {type: string}
                  mime_type: {type: [string, 'null']}
                  width_px: {type: [integer, 'null']}
                  height_px: {type: [integer, 'null']}
                  placement: {type: string, enum: [lead, gallery, inline]}
                  alt_text: {type: string}
                  caption: {type: [string, 'null']}
    TermRef:
      type: object
      required: [code, label]
      properties:
        code: {type: string}
        label: {type: string}
    EntityRef:
      type: object
      required: [public_id, label]
      properties:
        public_id: {type: string}
        label: {type: string}
        type: {$ref: '#/components/schemas/TermRef'}
...